requirements.txt
# NLP + LLM
spacy==3.8.5
sentence-transformers==4.1.0
torch>=1.12.0

# Resume Parsing
PyMuPDF==1.25.5

# Streamlit UI
streamlit==1.34.0

# Utilities
pandas==2.2.2
numpy==1.26.4

# (Optional, if doing fairness evaluation)
fairlearn==0.10.0
matplotlib==3.8.4
seaborn==0.13.2

# For config/env loading (if needed later)
python-dotenv==1.0.1
